Release 0.3.0

- Implements the monitoring tool setup
- Integration of the old SJG tool
- Ability to update pool registration information
- Code componentization ver3

+++ scripts/install.sh
@@ -7,8 +7,8 @@
 CNM_INST_DIR=/opt/cnm
 CNM_HOME=$HOME/cnm
 gum_version="0.14.5"
-# cnm_version="$(curl -s https://api.github.com/repos/btbf/sjg-tools/releases/latest | jq -r '.tag_name')"
-cnm_version="0.2.6"
+cnm_version="$(curl -s https://api.github.com/repos/btbf/sjg-tools/releases/latest | jq -r '.tag_name')"
+
 
 source ${HOME}/.bashrc
 
@@ -45,7 +45,10 @@ STAKE_VKEY_FILENAME="stake.vkey"
 STAKE_ADDR_FILENAME="stake.addr"
 STAKE_CERT_FILENAME="stake.cert"
 POOL_META_FILENAME="poolMetaData.json"
+POOL_ID_FILENAME="pool.id"
+POOL_ID_BECH32_FILENAME="pool.id-bech32"
 KOIOS_API="${3}"
+NODE_PROMETHEUS_PORT="12798"
 EOF
 }
 
@@ -63,16 +66,21 @@ EOF
 
 
 #ライブラリインストール
+
 echo "ライブラリをインストールします"
-sudo mkdir -p /etc/apt/keyrings
-curl -fsSL https://repo.charm.sh/apt/gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/charm.gpg
-echo "deb [signed-by=/etc/apt/keyrings/charm.gpg] https://repo.charm.sh/apt/ * *" | sudo tee /etc/apt/sources.list.d/charm.list
-sudo apt update && sudo apt install gum=${gum_version}
+if [ ! -e "/usr/bin/gum" ]; then
+    sudo mkdir -p /etc/apt/keyrings
+    curl -fsSL https://repo.charm.sh/apt/gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/charm.gpg
+    echo "deb [signed-by=/etc/apt/keyrings/charm.gpg] https://repo.charm.sh/apt/ * *" | sudo tee /etc/apt/sources.list.d/charm.list
+    sudo apt update && sudo apt install gum=${gum_version}
+fi
+sudo apt install git jq bc automake tmux rsync htop curl build-essential pkg-config libffi-dev libgmp-dev libssl-dev libtinfo-dev libsystemd-dev zlib1g-dev make g++ wget libncursesw5 libtool autoconf liblmdb-dev chrony fail2ban -y
 gum --version
 echo
 
 #CNODE Managerインストール
 echo "CNODE Managerをインストールします"
+mkdir -p $HOME/git
 cd $HOME/git || { echo "Failure"; exit 1; }
 wget -q https://github.com/btbf/sjg-tools/archive/refs/tags/${cnm_version}.tar.gz -O cnm.tar.gz
 tar xzvf cnm.tar.gz
@@ -99,7 +107,7 @@ if [ ! -d "${CNM_HOME}" ]; then
 
     if [ -d "${NODE_HOME}" ]; then echo -e "既存のネットワーク設定が見つかりました : ${NODE_CONFIG}\n";workDir=${NODE_HOME};syncNetwork=${NODE_CONFIG}; fi
 
-    nodeType=$(gum choose --header="セットアップノードタイプを選択して下さい" "ブロックプロデューサー" "リレー" --no-show-help)
+    nodeType=$(gum choose --header="セットアップノードタイプを選択して下さい" "ブロックプロデューサー" "リレー" "エアギャップ" --no-show-help)
     
     if [ ! -d "${NODE_HOME}" ]; then
         syncNetwork=$(gum choose --header="接続ネットワークを選択してください" --no-show-help "mainnet" "preview" "preprod" "Sancho-net")
@@ -112,25 +120,25 @@ if [ ! -d "${CNM_HOME}" ]; then
             NODE_CONFIG=mainnet
             NODE_NETWORK='"--mainnet"'
             CARDANO_NODE_NETWORK_ID=mainnet
-            koios_domain="https://api.koios.rest"
+            koios_domain="https://api.koios.rest/api/v1"
         ;;
         "preview" )
             NODE_CONFIG=preview
             NODE_NETWORK='"--testnet-magic 2"'
             CARDANO_NODE_NETWORK_ID=2
-            koios_domain="https://preview.koios.rest"
+            koios_domain="https://preview.koios.rest/api/v1"
         ;;
         "preprod" )
             NODE_CONFIG=preprod
             NODE_NETWORK='"--testnet-magic 1"'
             CARDANO_NODE_NETWORK_ID=1
-            koios_domain="https://preprod.koios.rest"
+            koios_domain="https://preprod.koios.rest/api/v1"
         ;;
         "Sancho-net" )
             NODE_CONFIG=sanchonet
             NODE_NETWORK='"--testnet-magic 4"'
             CARDANO_NODE_NETWORK_ID=4
-            koios_domain="https://sancho.koios.rest"
+            koios_domain="https://sancho.koios.rest/api/v1"
         ;;
     esac
 
@@ -154,7 +162,7 @@ if [ ! -d "${CNM_HOME}" ]; then
             echo export NODE_NETWORK="${NODE_NETWORK}" >> "${HOME}"/.bashrc
             echo export CARDANO_NODE_NETWORK_ID="${CARDANO_NODE_NETWORK_ID}" >> "${HOME}"/.bashrc
 
-            echo alias cnode='"journalctl -u cardano-node -f"' >> "${HOME}"/.bashrc
+            echo alias cnode='"sudo journalctl -u cardano-node -f"' >> "${HOME}"/.bashrc
             echo alias cnstart='"sudo systemctl start cardano-node"' >> "${HOME}"/.bashrc
             echo alias cnrestart='"sudo systemctl reload-or-restart cardano-node"' >> "${HOME}"/.bashrc
             echo alias cnstop='"sudo systemctl stop cardano-node"' >> "${HOME}"/.bashrc
